test(auto-reply): split response usage coverage
